feature: BoardDto 추가
