+++ src/main/kotlin/ui/input/InputUI.kt
@@ -1,4 +1,4 @@
-package ui
+package ui.input
 
 import domain.NumberOfMines
